Add Phases Return endpoint

+++ Source/Backend/Web/RoomReservation.Web.DataTransferModels/Room/ListedRoomResponseModel.cs
@@ -12,7 +12,7 @@ namespace RoomReservation.Web.DataTransferModels.Room
 
         public string ApartmentRoomNumber { get; set; }
 
-        public byte ResidentsCount { get; set; }
+        public int ResidentsCount { get; set; }
 
         public virtual void CreateMappings(Profile configuration)
         {
